fix 64 cpu bug.

+++ ZLPhotoLib/Classes/ZLPhotoBrowser/Views/ZLPhotoPickerBrowserPhotoScrollView.m
@@ -153,7 +153,7 @@
             }
 
             // 网络URL
-            [_photoImageView sd_setImageWithURL:photo.photoURL placeholderImage:thumbImage options:SDWebImageRetryFailed progress:^(long receivedSize, long expectedSize) {
+            [_photoImageView sd_setImageWithURL:photo.photoURL placeholderImage:thumbImage options:SDWebImageRetryFailed progress:^(NSInteger receivedSize, NSInteger expectedSize) {
                 [self setProgress:(double)receivedSize / expectedSize];
             } completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType, NSURL *imageURL) {
                 [self setProgress:1.0];
